chatbot/requirements.txt
langgraph>=0.0.15
langchain-cohere
langchain-core>=0.1.0
langchain-openai>=0.0.5
python-dotenv
python-dateutil>=2.8.2
pinecone[grpc]
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
